[testparser] Serialize empty test case list to empty array
Previously, empty "cases" lists in summary.json were sometimes `null`
and sometimes `[]` depending on the implementation of the corresponding
test parser. For example:
https://logs.chromium.org/logs/fuchsia/buildbucket/cr-buildbucket.appspot.com/8872159991166498784/+/steps/launch_collect/0/steps/0/0/steps/passed_tasks/0/steps/affected:QEMU__attempt_0_/0/logs/summary.json/0

It's very easy to mix up nil slices and empty slices in Go
and it generally doesn't matter which one you use, so instead of
changing the implementation of every language's test parser we can just
enforce in `Parse()` that we always return an empty slice instead of a
nil slice when no test cases were parsed.

The same thing is also necessary in tefmocheck, which is the other place
TestCaseResults are instantiated.

Change-Id: I0d98c7cdd231238eeb38500d03da95b8e1b0a8ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416334
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Shai Barack <shayba@google.com>