[swd-e2e-tests] Wait for a reboot if `update` fails
The `update` command has another reboot race, where it's possible that
after an OTA was successfully installed, the reboot process could tear down
the services that `update` was talking to before it receives notice that
the OTA successfully happened.

This patch changes the `Updater` service that upon error, wait 30 seconds
to see if the device reboots anyway. If it does, treat this as if the OTA
sucessfully installed and the device rebooted as normal. Otherwise, err
out.

Test: manually changed `update` command to always error, and observed the
test still succeeded. Replaced `update` with `exit 1`, to make sure it
errored out.

Fixed: 58086
Change-Id: Ie46bcca85cd34a4d0374c5fc478c1b8ea31920f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416358
Reviewed-by: Alex Legg <alexlegg@google.com>
Testability-Review: Alex Legg <alexlegg@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>