[fatfs] Fix shut down race.
When you shut down the execution scope, it can return before all futures
have actually been dropped, but we were relying on that to properly
clean up.  So, to address that, we stop shut down from relying on
everything to have been dropped and just flush instead.

I also took this opportunity to remove one of the levels of indirection
for the filesystem.  It was only necessary because the get_filesystem()
method returned Arc, but it doesn't need to.

Bug: 53796
Change-Id: If0bece3c27c6256aa32af3d79160c4b8430712ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416540
Reviewed-by: Simon Shields <simonshields@google.com>
Testability-Review: Simon Shields <simonshields@google.com>
Commit-Queue: Chris Suter <csuter@google.com>