[scenic] Refactor swapchain format selection.
This moves pipeline warmup to the time when swapchains are created
in order to not hard-code the image format.

Swapchain format is now determined by a single list of preferred
formats in the BufferPool class. The preferred format is still
B8G8R8A8 as before but adjusting it is as simple as changing the
order of formats in this list.

This can be extended to support a board specific preferred format
in the future if needed.

Bug: 42571
Test: fx test gfx_pixeltests gfx_unittests
Change-Id: I591b28afd4785167c429c237035a4868471c5c5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416197
Reviewed-by: Josh Gargus <jjosh@google.com>
Testability-Review: Josh Gargus <jjosh@google.com>
Commit-Queue: David Reveman <reveman@google.com>