[mdns] ensure query responses are sent immediately
This CL fixes a problem in which responses to PTR queries were sent a
second late, because outbound resource records weren't getting
promptly flushed out as messages.

'Agents' in the mdns service can 'send' resource records and
questions, but those items are accumulated in the Mdns class until
Mdns decides to 'SendMessages', which creates messages from the
accumulated items and transmits the messages. That happens at various
points that make sense for normal agent behavior. The problem arises
when an agent sends items asynchronously with respect to the times
when SendMessages is normally called. InstanceResponder, the agent in
charge of publishing service instances, does this when the client
responds to a GetPublication call.

This CL adds a protected method (FlushSentItems) that causes Mdns
to SendMessages. InstanceResponder is currently the only agent that
calls that method.

Test: fx test mdns_test
      Added a test that verifies the combined behavior of Mdns and
      InstanceResponder.

Fixed: 58141

Change-Id: I49ac0540027bd4757084a262f1bb4c49dd6dc020
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416838
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>