[ffx] add sdk target and enable cross-builds
In order to build for any SDK target for both x64 and arm64 we need
to enable cross-builds by way of using is_host instead of
current_toolchain comparison. This is still a little imprecise, as
really we only build for linux and mac targets today, but it's
approximately intent oriented instead, and simpler to spell.

Change-Id: I6fd57af0328d0bdb175cd8308818054aeed15441
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415482
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Craig Tiller <ctiller@google.com>
Commit-Queue: James Tucker <raggi@google.com>