[power_manager] Add ThermalShutdown node type
Creates a new node called ThermalShutdown. This node type will handle
the simple task of polling a given temperature sensor and checking the
temperature value against a configurable threshold. If the temperature
exceeds the threshold, the node will enforce a system reboot.

This functionality currently exists as part of the ThermalPolicy node,
but there are a couple reasons we need this in a standalone node:
  1) On some products we'd like to use a basic thermal policy like this
  without all of the extra functionality (and dependencies) that come
  along with the ThermalPolicy node. Currently we're forced to enable
  the entire ThermalPolicy node and use hacks to disable the
  functionality of that node that we don't want.

  2) We need the ability to monitor multiple temperature sensors for
  critical shutdown temperatures. The current ThermalPolicy isn't
  designed for this. By having the functionality brought into a
  standalone node, we can enable an instance of this new node for each
  temperature sensor we'd like to monitor.

Test: fx test power_manager_tests
Bug: 52898
Change-Id: I3608b01a729184f2f0632c83fb4a2db6d11bcd91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412626
Commit-Queue: Patrick Shickel <pshickel@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>
Testability-Review: Onath Claridge <claridge@google.com>