[build] Support for emitting GSYM files
This is not quite ready yet because llvm-gsymutil currently doesn't
correctly handle symbols with zero size, but it allows us to start
experimenting.

For now, we only support emitting GSYM for target-side ELF binaries.
We are going to expand this to host-side ELF and Mach-O binaries once
the tooling is stable.

Bug: 58294
Change-Id: I55b01264528dd24cb462fe04996346a9a488f3c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416214
Commit-Queue: Petr Hosek <phosek@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>