[power_manager] Use dynamic strings for node name
The Node trait defines a function `name()` that is intended to return
the human-readable name of the Node implementation, mainly for logging.
The function is currently defined as returning a `&'static str` but
this severely limits the flexibility that is sometimes required. For
example, the TemperatureHandler node returns "TemperatureHandler" as its
name, but it would be better described by the temperature driver that
it's bound to as well. With this change, the string returned by the
`name()` trait method is a dynamic String that allows the Node
implementation to use a more detailed name that wouldn't be possible
using a static string.

Test: fx test power_manager_tests
Change-Id: I714aad945526dcfd481086f654d309a16c4ae4d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413397
Commit-Queue: Patrick Shickel <pshickel@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>
Testability-Review: Onath Claridge <claridge@google.com>