Reland "[Settings] Use ExternalError when handler calls fail"
Fixed: 58214
Test: fx test setui_service_tests setui_client_interface_test

This reverts commit 671b19bdded6fa1b20302bc2543cf56f01f81586.

Reason for revert: Relanding after fixes.

There was a line that changed self.stored_stream that was moved up above
the set_volume and set_mute calls in set_volume in order to ensure it
still got set even if the call failed. It was causing the
stored volume level to be the same as the new volume level, and it
was not detecting the volume as changed after the first change. This
fixes it so that the stored stream is saved after comparison but before
the errors.

Also adds a test to make sure consecutive volume changes are registered.

Original change's description:
> Revert "[Settings] Use ExternalError when handler calls fail"
>
> This reverts commit 73518d0eadd4c90cda67fba76eeea9bb47e0c06e.
>
> Reason for revert: Breaks volume control
>
> Original change's description:
> > [Settings] Use ExternalError when handler calls fail
> >
> > This ensures that ControllerErrors are returned in the controllers' handle trait methods.
> >
> > Fixed: 57172
> > Test: fx test setui_service_tests
> >
> > Change-Id: Ib3df609469c39157b7a66214b3ba3c3911ad525d
> > Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414717
> > Commit-Queue: Chip Fukuhara <cfukuhara@google.com>
> > Reviewed-by: Paul Faria <paulfaria@google.com>
> > Reviewed-by: William Xiao <wxyz@google.com>
> > Testability-Review: William Xiao <wxyz@google.com>
>
> TBR=wxyz@google.com,cfukuhara@google.com,paulfaria@google.com,setui-review@google.com
>
> Change-Id: I5c4808e369b196322b3313d42abb1d9346644eb9
> No-Presubmit: true
> No-Tree-Checks: true
> No-Try: true
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416394
> Reviewed-by: Bryce Lee <brycelee@google.com>
> Commit-Queue: Bryce Lee <brycelee@google.com>

TBR=brycelee@google.com,wxyz@google.com,cfukuhara@google.com,paulfaria@google.com,setui-review@google.com

Change-Id: Ibbdb1acd1d03d69c08c9782b5ebe1510ea9690ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416615
Reviewed-by: William Xiao <wxyz@google.com>
Testability-Review: William Xiao <wxyz@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>