[wlan][brcmfmac] Only get wsec_key iovar for WEP tests in auth_test
The wsec_key iovar only succeeds when the security mode is WEP. It
does not make sense to check this iovar for other security modes,
such as WPA1 and WPA2 since the handshake to establish the key is not
simulated yet.

This change also removes the use of getting the 'wsec_key' iovar using
IovarsGet() which is the final use of that function outside of
sim_fw.cc and iovar_test.cc which are the only places where it should
be used.

Test: fx test auth_test

Change-Id: If935dde78c4f66da90c6052dbfa56f639308bd85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416707
Commit-Queue: Charles Celerier <chcl@google.com>
Testability-Review: Karthik Krishnaswami <karthikrish@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>