[brcmfmac] Wrap wlanif histograms in feature flag.
This change also reduces the verbosity of the logging when an attempt to
retrieve the detailed histogram fails, as requested in fxb/58065.

Test: a new test that disables the detailed histogram functionality
passes, showing that the stats query will continue to work when the
feature is disabled; existing brcmfmac-sim-test tests that exercise the
detailed histogram functionality continue to pass.

Bug: 58065
Change-Id: Ic59ac7eff3c0b50899fd2ae8f4046c56a03a4c27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416680
Reviewed-by: Joshua Conner <joshconner@google.com>
Testability-Review: Joshua Conner <joshconner@google.com>
Commit-Queue: Karl Ward <karlward@google.com>