[ot-stack][test] Removing flaky part of a test
Removing unnecessary part from TEST(Alarm, RemainingTimeBothAlarms).
The test intent is to set both micro-sec and milli-sec alarms
and to sleep for some time, and check remaining time on micro-sec
alarm to make sure it's correctly reported.

Since in infra, the test may sleep longer than programmed, it may result
in more possibilities to happen rather than intended case. The test detected
some of those scenarios and then checked the results against it.
However, there are many different such scenarios for this test and taking
care of all those adds to unnecessary complications and resulting in flakes as well.

The test intent is still met. The commented out part of the test - that
is making sure that remaining time is reported as 0 when alarm as fired
is already tested by other tests (clearmicrosectest and clearmillisectest),
so we're not loosing coverage

MULTIPLY: threadstack-unittests: 200
Test: fx test threadstack-unittests
Bug: 58032

Change-Id: I93b303d78cc4cdddc06f0949d95b3b8dc94b1992
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416795
Commit-Queue: Parag Dixit <paragdixit@google.com>
Reviewed-by: Parag Dixit <paragdixit@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Testability-Review: Jiaming (Charlie) Wang <jiamingw@google.com>