[aml][pwm] Add ability to protect pwm channels
Some pwm channels may be in use by higher ELs or be running
in a state configured by bootloader which is required to maintain
safe operating conditions.  This CL adds ability to mark
channels as protected in the board file.  This flag is checked
by the pwm driver (and protocol) to block initialization of and
access to channels marked as protected.

TEST: Tested on astro while monitoring PP800_EE rail
Fixed: 58137

Change-Id: I90cf5acc2e5a4fd99f3890e64ef83d6675470e91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416819
Testability-Review: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Eric Holland <hollande@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>