[camera][bin] Split factory component into CLI and Server
This CL pulls out the CLI portion of the FactoryServer component into
it's own component: camera-factory-cli. A future CL may add these two to
the same package and remove them from the main camera build tree.

To accomplish this, the original camera-factory component's config is provided
to sysmgr through the `config_data` mechanism. This target needs to be included
in product configs to take effect.

The CLI component opens a channel to both the
fuchsia.factory.camera Controller as well as the ISP, sending a command
to the appropriate target as is appropriate. All mentions of the ISP
have been removed from the original component, it is only responsible
for managing streams/capture. The CL also cleans up the CLI code for readability.

Additionally, a GetHandler function has been added to the FactoryServer
class along with a corresponding AddPublicService call in the main.

Test: Ran ISP commands through the CLI manually.
Bug: 58025
Bug: 58319
Bug: 58323

Change-Id: I06a29d68e31b5d2cb97df42625ff0d5be09ebebb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416559
Commit-Queue: Nic Zoghb <nzo@google.com>
Testability-Review: Nic Zoghb <nzo@google.com>
Reviewed-by: Ed James <edjames@google.com>