[kernel][vm] Remove non-const VmPageList methods
These non-const methods allow for modifying the markers in place, which
is generally an unsafe operation as it does not allow the page list to
cleanup empty nodes or perform other modifications.

In the occasional scenario where in place modification is desirable the
RemovePages method exists and is the intended method to be used.

Change-Id: I76c7a90e1519b9f4bc1c31eb988c5da0981a27cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416535
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>