[tap] Add doc.go and fix typos
When I first encountered this package I was very confused by the name. I
assumed it referred to TAP, the Google internal testing system, but it
actually refers to the Test Anything Protocol, which is open and not
affiliated with Google. So I added a doc.go to clarify.

Change-Id: If13e423c0698327ddb8464a9c1cfd1530e127406
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417780
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>