[bt][a2dp-sink] Add peer session setup
Move session setup out of ConnectedPeers so that the streams and tasks
related to the A2DP peer are started on first connect.  Move AVRCP task
to the peer session generation function.

This allows us to separate audio session creation from the AVRCP task,
and pass session ids to the stream media sink tasks, allowing us to tie
the streams that are generated to the media session for the peer.

Fixed: 42976
Bug: 43545

Test: updated tests in bt-a2dp-sink-tests
Change-Id: I0b91e61ccf1227d38f087495c5b482060b74bb37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417894
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>