[brcmfmac] Set bounds on error injection buffer
When specifying a replacement buffer, instead of just storing a
raw pointer to the replacement value specify a pointer to a vector
so that we can place a proper bound on the values that are
accessed at runtime.

Also add a test that verifies proper behavior for various inject
buffer lengths (greater than, equal to, and less than the expected
length of the output structure).

Bug: 51987
Test: Included (err_inject_test)
Change-Id: Iabbb673eb858066692b4b83775c7134473ca8ec4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417886
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Testability-Review: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Joshua Conner <joshconner@google.com>