[debugger] Update the module cache on init.
Updates the module cache as soon as the dl_debug_addr is known.
Previously the way we executed when launching a new process updated the
module list, but not when attaching to an existing process. This meant
the module cache was out-of-date.

Update the module cache when requested by the client. This shouldn't be
necessary but allows the user to "fix" future issues along these lines
where the list of modules is out-of-date, and ensures the frontend and
the backend are in-sync.

Change-Id: I0fec0ddcee60730e3cbb54961ddb19384f05af1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418314
Reviewed-by: Wez . <wez@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>