[time] Preliminary time.external implementation for roughtime
Adds a basic time.external.PushSource implementation for roughtime
with only WatchSample implemented.
In a PushSource implementation, the implementation decides when it
should poll for a new time sample. For this implementation, the
time source only schedules a poll for a new sample if it has an active
WatchSample request. The advantage of this approach is that it does
not result in unused polls while the deprecatedtimezone protocol is
still in use.
The implementation currently assumes there is only a single client
and uses a very simple retry strategy in which time samples are polled
again a second after a failure, or 30 minutes after a successful poll.
Note that unlike the deprecatedtimezone implementation, this does not
update system time by itself - it only reports what it believes the
time is.
Up next: Implement WatchStatus, update the retry strategy to use the
exponential backoff with cap strategy currently in Timekeeper.

Bug: 56833

Testing: fx test network-time-service-test
Change-Id: I48aaa7cff0f0ca10d79910282fe7df760635106c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415832
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Testability-Review: Jody Sankey <jsankey@google.com>