[time] Remove redundant log
Timekeeper and network_time_service both output logs when an attempt to
retrieve time is made. This removes the log in network_time_service
which indicates an attempt has started, and reduces the severity of a
log that is output repeatedly in case internet is not available.

Bug: 58390

Change-Id: I75d5c67a7de6c14e2b60b01952077a4f2f5ea152
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417997
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Testability-Review: Jody Sankey <jsankey@google.com>