[timekeeper] Move away from the ergonomic inspect API.
As discussed with the diagnostics team, Ergonomic inspect wasn't a good
fit for our use case since most of the data isn't already being
maintained for operational use.

This CL moves to using the raw inspect API with a mechanism for
exporting data structs that is amenable to automatic generation via
procedural macro in the future once we're happy we've got the correct
paradigm.

The end result seems fairly ergonomic and (excluding the
auto-generatable code) its also not too verbose. At the moment the main
benefit is we no longer need to keep two copies of everything - one
inside the inspect property and one inside the IValue. We retain the
nice feature of ergonomic inspect that data is never exported before its
initialized.  This CL doesn't change the structure of the output or the
tests.

Bug: 56871

Change-Id: Idef2f367382dfa735ea3ac931a3f253ef34679d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417896
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>