[storage][volume_image]: Add decompressor.
In order to check validity of the the sparse image header, the data
needs to be decompressed so the total length can be retrieved, in order
to prevent any sort of truncated file that might look fine, but extent
descriptors with out bound pointers.

Move LZ4result to its own header, so it can be shared between compressor
and decompressor.

TEST=storage-volume-image

Change-Id: Ib5e760f960fe31b138a9581aa7d1d1ece38af9f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416401
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>