[time] WatchStatus implementation for roughtime time source
Adds a basic WatchStatus implementation for roughtime that piggybacks
on the update loop for WatchSample and declares whether or not it is
healthy based on whether the last time sample succeeded.
Also updates SampleWatcher as a template so that it can be reused for
watching Status as well.

Bug: 56833

Testing: fx test network-time-service

Change-Id: I1c788b23dba3b5cc7df5a8447c9682b03908862f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414383
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Testability-Review: Jody Sankey <jsankey@google.com>