[fs_test] Support blobfs
Adds blobfs as a supported variant for the fs_test framework.

Blobfs is intentionally not added to fs_test::AllTestFilesystems, since
it is not expected to run as part of the regular fs_test suite. Instead,
blobfs will run its own integration suite, using the fixtures provided
by the fs_test library. (Blobfs is not a general-purpose filesystem,
which is what the fs_test suite is designed for.)

Tested: Built and ran a test component in a child CL.
Change-Id: I5b54d112318d2f33c7dc1a56f866de1b835cc5b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417738
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Testability-Review: Chris Suter <csuter@google.com>