[msd-vsi-vip] Fix WaitUntilIdle().
It appears possible for IsIdle() to become false for a brief
period after returning true. This results in occasional test
flakes.

This change modifies WaitUntilIdle() to return after the first
true result of IsIdle(), which matches the implementation of
the reference drivers.

Ran the tests for half an hour and the flake appears to be fixed.

BUG=58070

Change-Id: Ia056e3f82d2703c6e73644b1f2f5154d3147d60d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418536
Reviewed-by: Craig Stout <cstout@google.com>
Testability-Review: Craig Stout <cstout@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>