[flatland] Add a queueing interface for UberStructSystem.
The UberStructSystem::UberStructQueue class is the intermediary
between Flatland instances and the UberStructSystem.

Flatlnd instances queue UberStructs for session updates through an
UberStructQueue object. The UberStructSystem pops these UberStructs
when told to perform session updates.

At the moment, this class wraps a mutex and queue, but in the future
it could be entirely replaced with a lock-free queue entirely. Issue
57745 tracks progress on this.

Because the UberStructQueue is managed by a shared_ptr, the
UberStructSystem can also detect Flatland instance destruction by the
reference count on the shared_ptr and de-allocate resources
appropriately, rather than depending on the Flatland instance for
this signal, which would introduce the need for a lock on the map
that holds the UberStructQueues.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Change-Id: Ie6d465b6381f5279812e268dad92ff6071afc3fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414115
Reviewed-by: Mikael Pessa <mikaelpessa@google.com>
Testability-Review: Mikael Pessa <mikaelpessa@google.com>
Commit-Queue: Evan Birenbaum <birenbaum@google.com>