[camera][imx355] Initial imx355 driver
This CL contains the initial version of the IMX355 driver.
When tested on a luis platform board with the enabling CLs to follow,
the IMX355 responded to I2C requests and, when streaming was enabled,
produced frames as seen through the incrementing frame counter.

Change-Id: I1f9d1274466591172f6f1279cf694c9bd14ead8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417622
Reviewed-by: Nic Zoghb <nzo@google.com>
Testability-Review: Nic Zoghb <nzo@google.com>
Commit-Queue: John Sasinowski <jsasinowski@google.com>