[gidl] pin generated dart code to 2.8
This comment will tell the dart compile to compile this code using
version 2.8. This is required because this generated code does not
belong to a package in any package_config.json file so the compiler
will assume it should be compiled against the latest version. When
we turn on nullability this will fail to compile since it is not
converted to null-safety.

Change-Id: If37c9e613e0547cd3634ef972069d7e20dc513e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418277
Reviewed-by: Ian McKellar <ianloic@google.com>
Testability-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>