[e2e] Also test for time zone *changes*
The dart runner should respond to time zone changes, as well as
time zone initialization.

This change modifies the timezone e2e test to try reading the timezone,
then change the timezone again, and ensure that the clocks match up
inside the dart VM and outside.

Previous test tried to do this only once, and didn't test repeated
attempts.  This works for the "deprecatedtimezone" implementation,
because it does not consult the time zone when computing the time
offset. But the "intl" implementation is sensitive to this, since it
asks about the timezone to compute its offset.

There are two tests for this feature; however, for the time being one
is `#[ignored]` because the two tests seem to step on each other while
changing the global mutable state (the time zone).

Multiply: timezone-test: 50
Bug: 42387, 58383
Change-Id: I05ee9562ba0d8f3200cd13408b17998e52f3b3f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419497
Testability-Review: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>