[power_manager] Set up standalone Cobalt Metrics
The ThermalPolicy node currently contains all of the logic surrounding
Cobalt metrics reporting.  This has been fine so far, but became a
problem as soon as a second node was added that also contributes to the
Cobalt metrics "state" (that node is ThermalShutdown).

This change pulls the majority of the Cobalt metrics logic out of the
ThermalPolicy node and puts it into a standalone singleton instance. The
change also sets up the ThermalShutdown node to call into this new
Cobalt structure when the device reaches thermal shutdown.

Test: fx test power_manager_tests
Change-Id: I78a658ac8e23dca667da40cb100303fead19bd59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416881
Commit-Queue: Patrick Shickel <pshickel@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>
Testability-Review: Onath Claridge <claridge@google.com>