[kernel][x86][kasan] Check out-of-bounds accesses to globals
This CL instruments accesses to global variables in zircon.

The compiler adds a redzone to the right of every global object. Our
instrumentation poisons the redzones and in the steady-state traps on
accesses to the redzones to detect and panic on out of bounds accesses.

To enable this instrumentation, we add all kernel sections that can
have globals to the regions kASAN tracks; we also record every
contiguous region of globals in early boot and register them with
asan (posion their right redzones) in later boot.

Pair-programmed-with: mvanotti@
Bug: 30033 Kernelmode ASAN for Zircon
Change-Id: Iab3227f784fda32f5b8bf2405b304ce18bea432f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392703
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>