[go] Add go.mod in subdirectories containing Go code
This is an alternative to having a single root go.mod, which doesn't
work due to the Go toolchain being built as part of the Fuchsia build
process. When the Go toolchain make.bash runs, there's eventually a check
to ensure that it is not inside a go module.

This is a bit tricky as:
- /garnet depends on /src.
- /src depends on /garnet and /tools.
- /tools depends on /src and /garnet.

A follow up will include "replace" statements to make this dependency tree work.
This can't easily be added now because a valid version containing a go.mod is
required.

To be extra clear: this does not affect the GN build at all, since all go.mod
are not in the root directory. But this does help the IDEs a lot, making things
like goimports just work.

Bug: 57231
Change-Id: If099d666ed3ab9d253959677e44fbe64da4987d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410950
Testability-Review: Oliver Newman <olivernewman@google.com>
Commit-Queue: Marc-Antoine Ruel <maruel@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>