[fidl][go] Enable host FIDL conformance tests for go
Currently it is complicated to test potentially breaking FIDL changes in
go. The target needs to be booted to run the tests, but you can't boot if the
tests break FIDL. Therefore, you need to go back to the previous version
and boot and then patch in your new change to run tests.

This simplifies things. The FIDL conformance tests can run on host so
correctness can be checked before booting.

Test: fx test --host fidl_go_conformance_tests
Change-Id: I4a3a0ef693babe22a42eac8aca5c50f83baf2084
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410261
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>